fix 208. Properly handle unix time value
